requirements.txt
# Core dependencies
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# Vector database and embeddings
qdrant-client==1.7.0
sentence-transformers==2.2.2

# OCR and image processing
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3

# Google Gemini API
google-generativeai==0.3.2

# Utilities
tqdm==4.66.1
pathlib2==2.3.7
sqlite3

# Development and testing
pytest==7.4.2
black==23.9.1
flake8==6.1.0
